+++ model/src/cfd.rs
@@ -4,6 +4,7 @@ use crate::hex_transaction;
 use crate::olivia;
 use crate::olivia::BitMexPriceEventId;
 use crate::payout_curve;
+use crate::rollover::RolloverParams;
 use crate::FeeAccount;
 use crate::FundingFee;
 use crate::FundingRate;
@@ -1268,43 +1269,6 @@ impl Cfd {
         }
 
         self
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct RolloverParams {
-    pub price: Price,
-    pub quantity: Usd,
-    pub leverage: Leverage,
-    pub refund_timelock: u32,
-    pub fee_rate: TxFeeRate,
-    pub fee_account: FeeAccount,
-    pub current_fee: FundingFee,
-}
-
-impl RolloverParams {
-    pub fn new(
-        price: Price,
-        quantity: Usd,
-        leverage: Leverage,
-        refund_timelock: u32,
-        fee_rate: TxFeeRate,
-        fee_account: FeeAccount,
-        current_fee: FundingFee,
-    ) -> Self {
-        Self {
-            price,
-            quantity,
-            leverage,
-            refund_timelock,
-            fee_rate,
-            fee_account,
-            current_fee,
-        }
-    }
-
-    pub fn funding_fee(&self) -> &FundingFee {
-        &self.current_fee
     }
 }
